requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
deprecated==1.2.14
ecdsa==0.19.0
fake-useragent==1.5.1
fastapi==0.112.1
fastapi-limiter==0.1.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-resources==6.4.3
itsdangerous==2.2.0
limits==3.13.0
mysqlclient==2.2.4
packaging==24.1
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.8
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.32
starlette==0.38.2
typing-extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
wrapt==1.16.0
